+++ packages/Actions/CodeExecution/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @memberjunction/code-execution
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/core@5.3.1
+- @memberjunction/global@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- @memberjunction/core@5.3.0
+- @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes