+++ packages/Angular/Explorer/explorer-app/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @memberjunction/ng-explorer-app
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/ng-bootstrap@5.3.1
+- @memberjunction/ng-auth-services@5.3.1
+- @memberjunction/ng-explorer-core@5.3.1
+- @memberjunction/ng-workspace-initializer@5.3.1
+- @memberjunction/core@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- @memberjunction/ng-explorer-core@5.3.0
+- @memberjunction/ng-bootstrap@5.3.0
+- @memberjunction/ng-workspace-initializer@5.3.0
+- @memberjunction/ng-auth-services@5.3.0
+- @memberjunction/core@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes