+++ packages/Angular/Generic/entity-relationship-diagram/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @memberjunction/ng-entity-relationship-diagram
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/core@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- @memberjunction/core@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes