+++ packages/Angular/Generic/entity-viewer/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @memberjunction/ng-entity-viewer
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/ng-export-service@5.3.1
+- @memberjunction/ng-filter-builder@5.3.1
+- @memberjunction/ng-shared-generic@5.3.1
+- @memberjunction/ng-timeline@5.3.1
+- @memberjunction/core@5.3.1
+- @memberjunction/core-entities@5.3.1
+- @memberjunction/export-engine@5.3.1
+- @memberjunction/global@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- 1692c53: Viewing System fixes for sorting and filtering. Memory manager SQL fix.
+- Updated dependencies [1692c53]
+  - @memberjunction/core-entities@5.3.0
+  - @memberjunction/ng-shared-generic@5.3.0
+  - @memberjunction/ng-export-service@5.3.0
+  - @memberjunction/ng-filter-builder@5.3.0
+  - @memberjunction/ng-timeline@5.3.0
+  - @memberjunction/core@5.3.0
+  - @memberjunction/export-engine@5.3.0
+  - @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes