+++ packages/Angular/Generic/file-storage/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Change Log - @memberjunction/ng-file-storage
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/ng-shared@5.3.1
+- @memberjunction/ng-container-directives@5.3.1
+- @memberjunction/ng-shared-generic@5.3.1
+- @memberjunction/graphql-dataprovider@5.3.1
+- @memberjunction/core@5.3.1
+- @memberjunction/core-entities@5.3.1
+- @memberjunction/global@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- Updated dependencies [a6aea29]
+- Updated dependencies [1692c53]
+  - @memberjunction/graphql-dataprovider@5.3.0
+  - @memberjunction/core-entities@5.3.0
+  - @memberjunction/ng-shared@5.3.0
+  - @memberjunction/ng-shared-generic@5.3.0
+  - @memberjunction/ng-container-directives@5.3.0
+  - @memberjunction/core@5.3.0
+  - @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes