+++ packages/Angular/Generic/timeline/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @memberjunction/ng-timeline
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/core@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- @memberjunction/core@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes