+++ packages/Communication/entity-comm-server/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log - @memberjunction/entity-communications-server
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/communication-types@5.3.1
+- @memberjunction/communication-engine@5.3.1
+- @memberjunction/entity-communications-base@5.3.1
+- @memberjunction/core@5.3.1
+- @memberjunction/core-entities@5.3.1
+- @memberjunction/global@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- Updated dependencies [1692c53]
+  - @memberjunction/core-entities@5.3.0
+  - @memberjunction/communication-types@5.3.0
+  - @memberjunction/communication-engine@5.3.0
+  - @memberjunction/entity-communications-base@5.3.0
+  - @memberjunction/core@5.3.0
+  - @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes