+++ packages/ComponentRegistryClientSDK/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @memberjunction/component-registry-client-sdk
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/interactive-component-types@5.3.1
+- @memberjunction/core@5.3.1
+- @memberjunction/global@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- @memberjunction/interactive-component-types@5.3.0
+- @memberjunction/core@5.3.0
+- @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes