+++ packages/SQLServerDataProvider/CHANGELOG.md
@@ -1,5 +1,38 @@
 # Change Log - @memberjunction/sqlserver-dataprovider
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/ai@5.3.1
+- @memberjunction/aiengine@5.3.1
+- @memberjunction/ai-provider-bundle@5.3.1
+- @memberjunction/ai-vector-dupe@5.3.1
+- @memberjunction/actions-base@5.3.1
+- @memberjunction/actions@5.3.1
+- @memberjunction/encryption@5.3.1
+- @memberjunction/core@5.3.1
+- @memberjunction/core-entities@5.3.1
+- @memberjunction/global@5.3.1
+- @memberjunction/queue@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- Updated dependencies [1692c53]
+  - @memberjunction/core-entities@5.3.0
+  - @memberjunction/aiengine@5.3.0
+  - @memberjunction/ai-vector-dupe@5.3.0
+  - @memberjunction/actions-base@5.3.0
+  - @memberjunction/actions@5.3.0
+  - @memberjunction/encryption@5.3.0
+  - @memberjunction/queue@5.3.0
+  - @memberjunction/ai-provider-bundle@5.3.0
+  - @memberjunction/ai@5.3.0
+  - @memberjunction/core@5.3.0
+  - @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Minor Changes