+++ packages/Scheduling/engine/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @memberjunction/scheduling-engine
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/ai-agents@5.3.1
+- @memberjunction/ai-core-plus@5.3.1
+- @memberjunction/actions-base@5.3.1
+- @memberjunction/actions@5.3.1
+- @memberjunction/core@5.3.1
+- @memberjunction/core-entities@5.3.1
+- @memberjunction/global@5.3.1
+- @memberjunction/sqlserver-dataprovider@5.3.1
+- @memberjunction/scheduling-engine-base@5.3.1
+- @memberjunction/scheduling-base-types@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- Updated dependencies [ebf057a]
+- Updated dependencies [1692c53]
+  - @memberjunction/ai-agents@5.3.0
+  - @memberjunction/core-entities@5.3.0
+  - @memberjunction/ai-core-plus@5.3.0
+  - @memberjunction/actions-base@5.3.0
+  - @memberjunction/actions@5.3.0
+  - @memberjunction/sqlserver-dataprovider@5.3.0
+  - @memberjunction/scheduling-engine-base@5.3.0
+  - @memberjunction/core@5.3.0
+  - @memberjunction/global@5.3.0
+  - @memberjunction/scheduling-base-types@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes