+++ packages/VersionHistory/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @memberjunction/version-history
 
+## 5.3.1
+
+### Patch Changes
+
+- @memberjunction/core@5.3.1
+- @memberjunction/core-entities@5.3.1
+- @memberjunction/global@5.3.1
+
+## 5.3.0
+
+### Patch Changes
+
+- Updated dependencies [1692c53]
+  - @memberjunction/core-entities@5.3.0
+  - @memberjunction/core@5.3.0
+  - @memberjunction/global@5.3.0
+
 ## 5.2.0
 
 ### Patch Changes